Split formatter module into formatter/ package
