+++ a816/formatter/__init__.py
@@ -1,0 +1,8 @@
+"""`a816.formatter`: AST → formatted source. Public API re-exports."""
+
+from __future__ import annotations
+
+from a816.formatter.core import A816Formatter
+from a816.formatter.options import FormattingOptions
+
+__all__ = ["A816Formatter", "FormattingOptions"]
